Add getting income by id

+++ src/main/java/ru/mityugov/budget_planner/repository/IncomeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.mityugov.budget_planner.repository.entity.IncomeEntity;
 
 @Repository
-public interface IncomeRepository extends JpaRepository<IncomeEntity, Integer> {
+public interface IncomeRepository extends JpaRepository<IncomeEntity, Long> {
 }
